requirements.txt
Django>=4.1.1,<4.2
djangorestframework>=3.13.1,<3.14
djangorestframework-simplejwt==5.2.2
django-extensions==3.2.3
drf-yasg==1.21.6
django-cors-headers==4.1.0
factory-boy==3.2.1